fourteenth commit

+++ src/users/usuario.repository.ts
@@ -33,11 +33,21 @@ export class UsuarioRepository {
   }
 
   async atualiza(id: string, dadosDeAtualizacao: Partial<UsuarioEntity>) {
+    if (!id) {
+      console.error("ID está vazio ou nulo");
+      throw new Error('ID está vazio ou nulo');
+    }
+
     await this.userRepository.update(id, dadosDeAtualizacao);
     return await this.buscaPorId(id);
   }
 
   async remove(id: string) {
+    if (!id) {
+      console.error("ID está vazio ou nulo");
+      throw new Error('ID está vazio ou nulo');
+    }
+
     const usuarioRemovido = await this.buscaPorId(id);
     await this.userRepository.delete(id);
     return usuarioRemovido;
